+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 # Changelog
 
+# 10.1.2
+
+ * Express v1.0 firmware updates.
+ * Regimen Farm Event scheduler fixes.
+
 # 10.1.1
 
- * Genesis v1.5 and Express v1.0 firmware updates.
  * Changes to TTY to allow use of DIY boards
  * Upgrade underlying OS for RPi3 Nerves system (`ERL_CRASH_DUMP_SECONDS=-1`)